+++ src/components/Main/Articles/blockNotes/Note.jsx
@@ -1,26 +1,95 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
+import PropTypes from 'prop-types';
 
+import { Avatar } from 'antd';
 import openNote from '../icons/open_note.svg';
 import closeNote from '../icons/close_note.svg';
-
-const TextOne = `Dolor sit amet, consectetur adipiscing elit. Maecenas sed congue urna. Etiam in pretium
-orci. Nunc id nulla sit amet ex lacinia euismod a eu risus. Sed consequat viverra libero
-congue vestibulum. Ut dolor odio, posuere vitae accumsan a, placerat ut odio. Proin in lorem
-dui. Maecenas viverra convallis nisi eget tincidunt. Vivamus nec tempus nunc, sit amet
-auctor massa. Morbi lobortis, elit eu malesuada malesuada, quam mauris semper turpis, vitae
-dapibus erat nibh at purus. Ut id odio massa.`;
-
-const TextTwo = `Nulla in metus dictum, dapibus justo sit amet, tristique purus. Pellentesque blandit nibh at
-leo venenatis, in semper ipsum dictum. Nulla porttitor ligula quam, quis auctor felis
-consectetur non. Praesent at auctor metus. Etiam molestie tincidunt justo eget laoreet.
-Quisque vestibulum pretium tortor in mattis. Nullam nec elit vitae tellus interdum vehicula.
-Orci varius natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus.
-Praesent congue arcu at accumsan luctus. Nam sed diam nibh. In at lectus sodales, interdum
-turpis at, semper nisl.`;
+import actionAddNone from '../icons/action_addNote.svg';
+import actionLike from '../icons/action_like.svg';
+import actionComment from '../icons/action_comment.svg';
+import actionRepost from '../icons/action_repost.svg';
+import BlockComments from '../blockComment/BlockComments';
 
 const Wrapper = styled.div`
   position: relative;
+  border-bottom: 1px solid #515151;
+`;
+
+const WrapperNote = styled.div`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin: 30px 0;
+`;
+
+const UserInfo = styled.div`
+  display: flex;
+  justify-content: flex-start;
+  align-items: center;
+`;
+
+const TitleWrap = styled.div`
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: flex-start;
+  margin-left: 25px;
+`;
+const Title = styled.h3`
+  font-weight: 500;
+  font-size: 16px;
+  line-height: 20px;
+  margin: 0 0 5px 0;
+  color: black;
+`;
+const SubtitleTime = styled.span`
+  font-weight: 500;
+  font-size: 13px;
+  line-height: 16px;
+  color: #515151;
+`;
+
+const UserActions = styled.div`
+  display: flex;
+  justify-content: flex-start;
+  align-items: center;
+`;
+const UserAction = styled.div`
+  width: 30px;
+  height: 30px;
+  background-position: center center;
+  background-repeat: no-repeat;
+  cursor: pointer;
+`;
+
+const Action = styled.div`
+  display: flex;
+  justify-content: flex-start;
+  align-items: center;
+  &:not(:last-child) {
+    margin-right: 45px;
+  }
+`;
+
+const ActionAddNote = styled(UserAction)`
+  background-image: url(${actionAddNone});
+`;
+const ActionLike = styled(UserAction)`
+  background-image: url(${actionLike});
+`;
+const ActionComment = styled(UserAction)`
+  background-image: url(${actionComment});
+`;
+const ActionRepost = styled(UserAction)`
+  background-image: url(${actionRepost});
+`;
+
+const CountAction = styled.span`
+  margin-left: 10px;
+  font-size: 18px;
+  line-height: 160.9%;
+  color: black;
 `;
 
 const TitleText = styled.p`
@@ -39,8 +108,6 @@ const Text = styled.div`
   margin: 0 0 40px 0;
   color: #000000;
   text-align: start;
-  //height: 100px;
-  //height: 100%;
   height: ${(props) => props.height};
   overflow: hidden;
 `;
@@ -56,7 +123,7 @@ const BtnOpenNote = styled.button`
   cursor: pointer;
   border: none;
   position: absolute;
-  bottom: 0;
+  bottom: 50px;
   right: 0;
 
   &:focus {
@@ -70,7 +137,6 @@ const TagsList = styled.ul`
   justify-content: flex-start;
   align-items: center;
   margin: 45px 0 55px 0;
-  margin-top: 10px;
 `;
 const TagItem = styled.li`
   list-style: none;
@@ -81,29 +147,72 @@ const TagItem = styled.li`
   }
 `;
 
-const arrayTags = [
-  { id: 0, tag: '#nunc' },
-  { id: 1, tag: '#dolor' },
-];
-
-const Note = () => {
+const Note = ({ users }) => {
   const [height, setHeight] = useState(`${100}px`);
   const minHeight = '100px';
   const fullHeight = '100%';
+  const { first_name: firstName, last_name: lastName, avatarka, posts } = users.users;
+  const {
+    title,
+    text,
+    persist_date: persistDate,
+    bookmarks_count: bookmarksCount,
+    repost_count: repostCount,
+    tags,
+    likes_count: likesCount,
+    comments_count: commentsCount,
+    comments,
+  } = posts[0];
+  const fullName = `${firstName} ${lastName}`;
+
   return (
-    <Wrapper>
-      <TitleText>Lorem ipsum</TitleText>
-      <Text height={height}>{TextOne + TextTwo}</Text>
-      <BtnOpenNote
-        isOpen={height !== minHeight}
-        onClick={() => (height === minHeight ? setHeight(fullHeight) : setHeight(minHeight))}
-      />
-      <TagsList>
-        {arrayTags.map((item) => (
-          <TagItem key={item.id}>{item.tag}</TagItem>
-        ))}
-      </TagsList>
-    </Wrapper>
+    <>
+      <WrapperNote>
+        <UserInfo>
+          <Avatar src={avatarka} />
+          <TitleWrap>
+            <Title>{fullName}</Title>
+            <SubtitleTime>{persistDate}</SubtitleTime>
+          </TitleWrap>
+        </UserInfo>
+        <UserActions>
+          <Action>
+            <ActionAddNote />
+            <CountAction>{bookmarksCount}</CountAction>
+          </Action>
+          <Action>
+            <ActionLike />
+            <CountAction>{likesCount}</CountAction>
+          </Action>
+          <Action>
+            <ActionComment />
+            <CountAction>{commentsCount}</CountAction>
+          </Action>
+          <Action>
+            <ActionRepost />
+            <CountAction>{repostCount}</CountAction>
+          </Action>
+        </UserActions>
+      </WrapperNote>
+      <Wrapper>
+        <TitleText>{title}</TitleText>
+        <Text height={height}>{text}</Text>
+        <BtnOpenNote
+          isOpen={height !== minHeight}
+          onClick={() => (height === minHeight ? setHeight(fullHeight) : setHeight(minHeight))}
+        />
+        <TagsList>
+          {tags.map((item) => (
+            <TagItem key={item.id}>{item.tag}</TagItem>
+          ))}
+        </TagsList>
+      </Wrapper>
+      <BlockComments comments={comments} />
+    </>
   );
+};
+
+Note.propTypes = {
+  users: PropTypes.objectOf(PropTypes.object).isRequired,
 };
 export default Note;
